dont use int, use ssize_t

+++ itb_net.h
@@ -98,8 +98,8 @@ ITBDEF int itb_make_tcp();
 ITBDEF int itb_make_connected(const char *address, const char *port);
 ITBDEF int itb_accept_blind(int sfd);
 ITBDEF int itb_accept_addr(int sfd, struct sockaddr_storage *addr);
-ITBDEF int itb_recv(int sockfd, uint8_t *buffer, size_t len);
-ITBDEF int itb_send(int sockfd, const uint8_t *buffer, size_t len);
+ITBDEF ssize_t itb_recv(int sockfd, uint8_t *buffer, size_t len);
+ITBDEF ssize_t itb_send(int sockfd, const uint8_t *buffer, size_t len);
 
 //==>unix wrappers<==
 ITBDEF int itb_make_bound_unix(const char *path);
@@ -109,11 +109,11 @@ ITBDEF int itb_make_connected_unix(const char *path);
 //functions for setting up UDP
 ITBDEF int itb_make_bound_udp(int port);
 ITBDEF int itb_make_udp();
-ITBDEF int itb_read_message(int sockfd, uint8_t *buffer, size_t len);
-ITBDEF int itb_read_message_addr(
+ITBDEF ssize_t itb_read_message(int sockfd, uint8_t *buffer, size_t len);
+ITBDEF ssize_t itb_read_message_addr(
     int sockfd, uint8_t *buffer, size_t len, struct sockaddr_storage *addr);
-ITBDEF int itb_read_message_port(int sockfd, uint8_t *buffer, size_t len, int *port);
-ITBDEF int itb_send_message(
+ITBDEF ssize_t itb_read_message_port(int sockfd, uint8_t *buffer, size_t len, int *port);
+ITBDEF ssize_t itb_send_message(
     int sockfd, const uint8_t *buffer, size_t len, const struct sockaddr_storage *addr);
 
 //==>epoll wrappers<==
@@ -292,8 +292,8 @@ int itb_accept_addr(int sfd, struct sockaddr_storage *addr) {
     return ret;
 }
 
-int itb_recv(int sockfd, uint8_t *restrict buffer, size_t len) {
-    int total = 0, ret;
+ssize_t itb_recv(int sockfd, uint8_t *restrict buffer, size_t len) {
+    ssize_t total = 0, ret;
 readmsg:
     itb_ensure_nonblock((ret = recv(sockfd, buffer + total, len - total, 0)) != -1);
     if (ret <= 0)
@@ -302,8 +302,8 @@ readmsg:
     goto readmsg;
 }
 
-int itb_send(int sockfd, const uint8_t *restrict buffer, size_t len) {
-    int ret;
+ssize_t itb_send(int sockfd, const uint8_t *restrict buffer, size_t len) {
+    ssize_t ret;
     itb_ensure_nonblock((ret = send(sockfd, buffer, len, 0)) != -1);
     return ret;
 }
@@ -402,8 +402,8 @@ int itb_make_udp(void) {
     return sfd;
 }
 
-int itb_read_message(int sockfd, uint8_t *restrict buffer, size_t len) {
-    int total = 0, ret;
+ssize_t itb_read_message(int sockfd, uint8_t *restrict buffer, size_t len) {
+    ssize_t total = 0, ret;
 readmsg:
     itb_ensure_nonblock((ret = recvfrom(sockfd, buffer + total, len - total, 0, NULL, NULL)) != -1);
     if (ret == -1)
@@ -412,11 +412,11 @@ readmsg:
     goto readmsg;
 }
 
-int itb_read_message_port(int sockfd, uint8_t *restrict buffer, size_t len, int *restrict port) {
+ssize_t itb_read_message_port(int sockfd, uint8_t *restrict buffer, size_t len, int *restrict port) {
     struct sockaddr_storage addr;
     socklen_t addr_len;
     char hbuf[NI_MAXHOST], sbuf[NI_MAXSERV];
-    int total = 0, ret;
+    ssize_t total = 0, ret;
 readmsg:
     addr_len = sizeof(struct sockaddr_storage);
     itb_ensure_nonblock((ret = recvfrom(sockfd, buffer + total, len - total, 0,
@@ -432,10 +432,10 @@ readmsg:
     goto readmsg;
 }
 
-int itb_read_message_addr(
+ssize_t itb_read_message_addr(
     int sockfd, uint8_t *restrict buffer, size_t len, struct sockaddr_storage *addr) {
     socklen_t addr_len;
-    int total = 0, ret;
+    ssize_t total = 0, ret;
 readmsg:
     addr_len = sizeof(struct sockaddr_storage);
     itb_ensure_nonblock(
@@ -447,9 +447,9 @@ readmsg:
     goto readmsg;
 }
 
-int itb_send_message(int sockfd, const uint8_t *restrict buffer, size_t len,
+ssize_t itb_send_message(int sockfd, const uint8_t *restrict buffer, size_t len,
     const struct sockaddr_storage *restrict addr) {
-    int ret;
+    ssize_t ret;
     socklen_t addr_len = sizeof(struct sockaddr_storage);
     itb_ensure_nonblock(
         (ret = sendto(sockfd, buffer, len, 0, (struct sockaddr *)addr, addr_len)) != -1);
